kiss: fix bugs with pkg_build
Duplicates were not filtered out of the argument list as originally
intended due to the creation of $explicit2 disregarding $explicit.
The total package count is now aware of duplicates also.

The total package count (previously nexplicit) would also incorrectly
count packages which were removed from the explicit list. It would
silently /not prompt/ when a package went from explicit to implicit.

I have also updated the output to reflect whether a package is
explicit or implicit. ie, whether it's a dependency or not.