Make kiss update scriptable
Continue the upgrade even if kiss itself is upgrade instead of aborting
early. This is achieved by exec-ing into the freshly installed kiss.

The goal of this is to allow `kiss u` to be run as a script and to
always succeed in updating packages given the packages can actually be
accessed and built.