preserve installed file timestamps throughout kiss
install_files(): preserve the timestamps

This patch fixes packages which rely on using the timestamp of their
installed files to determine "newness", where if a file is newer, they
will automatically recompile it. This patch prevents those recompilations,
since the timestamps are now at their expected value. We only run `touch`
on non-symlink files, since if we run touch on symlinks, we can run into
file resolution errors, which will abort the installation of packages and
result in a dirty filesystem.

pkg_swap(): also preserve timestamps